Move Kafka-based repos into a separate package.
