+++ deployment/local-kafka/src/main/java/io/spine/examples/todolist/server/KafkaBoundedContextFactory.java
@@ -32,6 +32,9 @@ import io.spine.examples.todolist.context.BoundedContextFactory;
 import io.spine.examples.todolist.q.projection.DraftTasksView;
 import io.spine.examples.todolist.q.projection.LabelledTasksView;
 import io.spine.examples.todolist.q.projection.MyListView;
+import io.spine.examples.todolist.server.repository.LabelKAggregateRepository;
+import io.spine.examples.todolist.server.repository.TaskKRepository;
+import io.spine.examples.todolist.server.repository.TaskLabelsKRepository;
 import io.spine.server.BoundedContext;
 import io.spine.server.aggregate.AggregatePartRepository;
 import io.spine.server.aggregate.AggregateRepository;
